Complete get weather from specific time
This is a big change commit. Here's the changes that I made:
- Increase the size of weather icons in combobox from 32 to 50 so that
the users can see the weather icons in detail. I also made a lighter
background behind the combobox's items in order that the black parts of
weather icons aren't hidden in the original black background. The
function getWeatherIconUrlFromCode() in weather.cpp has 1 more param is
scale, By default the scale is 4, but with weather icons in
weatherTimeSelector combobox, the scale is 2

- The currentHourInDay is automactically changed to the time on you
system. (the time is rounded)

- You can now open up weatherTimeSelector combobox and choose the
specific time, the program will show up for you the information about
that time that you chose

- Some other changes aren't important, it doesn't affect much for our
program. That's just deleting my previous mistake code